refactor(WalletSignup): Moved some functionality to WalletNetwork, use of promise
